Re-factor of ui_json utils. Handle groupOptional
