+++ geoh5py/ui_json/validation.py
@@ -25,6 +25,7 @@ from geoh5py.groups import PropertyGroup
 from geoh5py.shared import Entity
 from geoh5py.shared.exceptions import RequiredValidationError
 from geoh5py.shared.validators import BaseValidator, TypeValidator
+from geoh5py.ui_json.utils import group_optional
 from geoh5py.workspace import Workspace
 
 
@@ -115,7 +116,6 @@ class InputValidation:
     @staticmethod
     def _validations_from_uijson(ui_json: dict[str, Any]) -> dict[str, dict]:
         """Determine base set of validations from ui.json structure."""
-
         validations: dict[str, dict] = {}
         for key, item in ui_json.items():
             if not isinstance(item, dict):
@@ -164,8 +164,9 @@ class InputValidation:
                     "types": [check_type],
                 }
 
-            if item.get("optional") and "types" in validations[key]:
-
+            if (
+                item.get("optional") or group_optional(ui_json, item.get("group", ""))
+            ) and "types" in validations[key]:
                 validations[key]["types"] += [type(None)]
 
         return validations
